Fix service for region url maps and add example config.
