+++ examples/compute/region-url-map.gyro
@@ -1,0 +1,92 @@
+google::bucket bucket-example-one
+    name: "bucket-example-one"
+    location: "US-CENTRAL1"
+end
+
+google::compute-health-check health-check-example
+    name: "health-check-example"
+    check-interval-sec: 30
+    description: "The description goes here."
+    healthy-threshold: 8
+    timeout-sec: 29
+    unhealthy-threshold: 6
+
+    http-health-check
+        proxy-header: "PROXY_V1"
+        request-path: "/myapp"
+        response: "okay"
+    end
+end
+
+google::compute-backend-bucket backend-bucket-example
+    name: "backend-bucket-example"
+    description: "Global backend bucket."
+    bucket: $(google::bucket bucket-example-one)
+end
+
+google::compute-network lb-ig-network-example
+    name: "lb-ig-network-example"
+    description: "lb-ig-network-example"
+    routing-mode: "Regional"
+end
+
+google::compute-instance-group lb-instance-group-example
+    name: "lb-instance-group-example"
+    description: "instance-group-example-description"
+    zone: "us-central1-b"
+    network: $(google::compute-network lb-ig-network-example)
+
+    named-port
+        name: "port-a"
+        port: 123
+    end
+
+    named-port
+        name: "port-b"
+        port: 490
+    end
+end
+
+google::compute-region-backend-service region-backend-service-example
+    name: 'region-backend-service-example'
+    region: "us-central1"
+    description: 'region-backend-service-example-desc'
+
+    backend
+        group: $(google::compute-instance-group lb-instance-group-example)
+        balancing-mode: "UTILIZATION"
+    end
+
+    health-check: [ $(google::compute-health-check health-check-example) ]
+
+    connection-draining
+        draining-timeout-sec: 41
+    end
+
+    load-balancing-scheme: "INTERNAL"
+
+    protocol: "HTTPS"
+    session-affinity: "NONE"
+end
+
+google::compute-region-url-map region-url-map-example
+    name: "region-url-map-example"
+    region: "us-central1"
+    description: "Region URL map description."
+    default-region-backend-service: $(google::compute-region-backend-service region-backend-service-example)
+
+    host-rule
+        hosts: [ "example.com" ]
+        path-matcher: "abc-path"
+    end
+
+    path-matcher
+        name: "abc-path"
+        default-region-backend-service: $(google::compute-region-backend-service region-backend-service-example)
+        path-rule
+            region-backend-service: $(google::compute-region-backend-service region-backend-service-example)
+            paths: [ "/abc" ]
+        end
+    end
+end
+
