+++ src/main/java/gyro/google/compute/UrlMapResource.java
@@ -16,15 +16,20 @@
 
 package gyro.google.compute;
 
+import java.util.ArrayList;
+import java.util.List;
 import java.util.Set;
 
 import com.google.api.services.compute.Compute;
 import com.google.api.services.compute.model.Operation;
 import com.google.api.services.compute.model.UrlMap;
+import gyro.core.GyroException;
 import gyro.core.GyroUI;
 import gyro.core.Type;
 import gyro.core.resource.Resource;
 import gyro.core.scope.State;
+import gyro.core.validation.ConflictsWith;
+import gyro.core.validation.ValidationError;
 
 /**
  * Creates a global URL map.
@@ -57,6 +62,67 @@ import gyro.core.scope.State;
 @Type("compute-url-map")
 public class UrlMapResource extends AbstractUrlMapResource {
 
+    private BackendBucketResource defaultBackendBucket;
+    private BackendServiceResource defaultBackendService;
+    private RegionBackendServiceResource defaultRegionBackendService;
+
+    /**
+     * The default backend bucket resource to which traffic is directed if none of the host rules match. Conflicts with ``default-backend-service`` and ``default-region-backend-service``.
+     */
+    @ConflictsWith({ "default-backend-service", "default-region-backend-service" })
+    public BackendBucketResource getDefaultBackendBucket() {
+        return defaultBackendBucket;
+    }
+
+    public void setDefaultBackendBucket(BackendBucketResource defaultBackendBucket) {
+        this.defaultBackendBucket = defaultBackendBucket;
+    }
+
+    /**
+     * The default backend service resource to which traffic is directed if none of the host rules match. Conflicts with ``default-backend-bucket`` and ``default-region-backend-service``.
+     */
+    @ConflictsWith({ "default-backend-bucket", "default-region-backend-service" })
+    public BackendServiceResource getDefaultBackendService() {
+        return defaultBackendService;
+    }
+
+    public void setDefaultBackendService(BackendServiceResource defaultBackendService) {
+        this.defaultBackendService = defaultBackendService;
+    }
+
+    /**
+     * The default region backend service resource to which traffic is directed if none of the host rules match. Conflicts with ``default-backend-bucket`` and ``default-backend-service``.
+     */
+    @ConflictsWith({ "default-backend-bucket", "default-backend-service" })
+    public RegionBackendServiceResource getDefaultRegionBackendService() {
+        return defaultRegionBackendService;
+    }
+
+    public void setDefaultRegionBackendService(RegionBackendServiceResource defaultRegionBackendService) {
+        this.defaultRegionBackendService = defaultRegionBackendService;
+    }
+
+    @Override
+    public void copyFrom(UrlMap urlMap) {
+        super.copyFrom(urlMap);
+
+        String defaultService = urlMap.getDefaultService();
+        setDefaultBackendBucket(null);
+        if (BackendBucketResource.parseBackendBucket(getProjectId(), defaultService) != null) {
+            setDefaultBackendBucket(findById(BackendBucketResource.class, defaultService));
+        }
+
+        setDefaultBackendService(null);
+        if (BackendServiceResource.parseBackendService(getProjectId(), defaultService) != null) {
+            setDefaultBackendService(findById(BackendServiceResource.class, defaultService));
+        }
+
+        setDefaultRegionBackendService(null);
+        if (RegionBackendServiceResource.parseRegionBackendService(getProjectId(), defaultService) != null) {
+            setDefaultRegionBackendService(findById(RegionBackendServiceResource.class, defaultService));
+        }
+    }
+
     @Override
     public boolean doRefresh() throws Exception {
         Compute client = createComputeClient();
@@ -68,7 +134,21 @@ public class UrlMapResource extends AbstractUrlMapResource {
     public void doCreate(GyroUI ui, State state) throws Exception {
         Compute client = createComputeClient();
 
-        Operation response = client.urlMaps().insert(getProjectId(), toUrlMap(null)).execute();
+        UrlMap urlMap = toUrlMap(null);
+        String defaultService;
+        if (getDefaultBackendBucket() != null) {
+            defaultService = getDefaultBackendBucket().getSelfLink();
+        } else if (getDefaultBackendService() != null) {
+            defaultService = getDefaultBackendService().getSelfLink();
+        } else if (getDefaultRegionBackendService() != null) {
+            defaultService = getDefaultRegionBackendService().getSelfLink();
+        } else {
+            throw new GyroException(
+                "Either 'default-backend-bucket', 'default-backend-service', or 'default-region-backend-service' is required!");
+        }
+        urlMap.setDefaultService(defaultService);
+
+        Operation response = client.urlMaps().insert(getProjectId(), urlMap).execute();
         waitForCompletion(client, response);
 
         refresh();
@@ -91,5 +171,19 @@ public class UrlMapResource extends AbstractUrlMapResource {
         Compute client = createComputeClient();
         Operation response = client.urlMaps().delete(getProjectId(), getName()).execute();
         waitForCompletion(client, response);
+    }
+
+    @Override
+    public List<ValidationError> validate() {
+        List<ValidationError> errors = new ArrayList<>();
+
+        if (getDefaultBackendBucket() == null && getDefaultBackendService() == null
+            && getDefaultRegionBackendService() == null) {
+            errors.add(new ValidationError(
+                this,
+                null,
+                "Either 'default-backend-bucket', 'default-backend-service', or 'default-region-backend-service' is required!"));
+        }
+        return errors;
     }
 }
